+++ esp-hal/src/dma/buffers.rs
@@ -4,24 +4,340 @@ use core::{
 };
 
 use super::*;
-use crate::soc::is_slice_in_dram;
-#[cfg(esp32s3)]
-use crate::soc::is_slice_in_psram;
+use crate::soc::{is_slice_in_dram, is_slice_in_psram};
+#[cfg(psram_dma)]
+use crate::soc::{is_valid_psram_address, is_valid_ram_address};
 
-/// Burst transfer configuration.
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+/// Error returned from Dma[Rx|Tx|RxTx]Buf operations.
+#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum BurstConfig {
-    /// Burst mode is disabled.
-    Disabled,
+pub enum DmaBufError {
+    /// The buffer is smaller than the requested size.
+    BufferTooSmall,
 
-    /// Burst mode is enabled.
-    Enabled,
+    /// More descriptors are needed for the buffer size.
+    InsufficientDescriptors,
+
+    /// Descriptors or buffers are not located in a supported memory region.
+    UnsupportedMemoryRegion,
+
+    /// Buffer address or size is not properly aligned.
+    InvalidAlignment(DmaAlignmentError),
+
+    /// Invalid chunk size: must be > 0 and <= 4095.
+    InvalidChunkSize,
+}
+
+/// DMA buffer alignment errors.
+#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum DmaAlignmentError {
+    /// Buffer address is not properly aligned.
+    Address,
+
+    /// Buffer size is not properly aligned.
+    Size,
+}
+
+impl From<DmaAlignmentError> for DmaBufError {
+    fn from(err: DmaAlignmentError) -> Self {
+        DmaBufError::InvalidAlignment(err)
+    }
+}
+
+cfg_if::cfg_if! {
+    if #[cfg(psram_dma)] {
+        /// Burst size used when transferring to and from external memory.
+        #[derive(Clone, Copy, PartialEq, Eq, Debug)]
+        #[cfg_attr(feature = "defmt", derive(defmt::Format))]
+        pub enum ExternalBurstConfig {
+            /// 16 bytes
+            Size16 = 16,
+
+            /// 32 bytes
+            Size32 = 32,
+
+            /// 64 bytes
+            Size64 = 64,
+        }
+
+        impl ExternalBurstConfig {
+            /// The default external memory burst length.
+            pub const DEFAULT: Self = Self::Size16;
+        }
+
+        impl Default for ExternalBurstConfig {
+            fn default() -> Self {
+                Self::DEFAULT
+            }
+        }
+
+        /// Internal memory access burst mode.
+        #[derive(Clone, Copy, PartialEq, Eq, Debug)]
+        #[cfg_attr(feature = "defmt", derive(defmt::Format))]
+        pub enum InternalBurstConfig {
+            /// Burst mode is disabled.
+            Disabled,
+
+            /// Burst mode is enabled.
+            Enabled,
+        }
+
+        impl InternalBurstConfig {
+            /// The default internal burst mode configuration.
+            pub const DEFAULT: Self = Self::Disabled;
+        }
+
+        impl Default for InternalBurstConfig {
+            fn default() -> Self {
+                Self::DEFAULT
+            }
+        }
+
+        /// Burst transfer configuration.
+        #[derive(Clone, Copy, PartialEq, Eq, Debug)]
+        #[cfg_attr(feature = "defmt", derive(defmt::Format))]
+        pub struct BurstConfig {
+            /// Configures the burst size for PSRAM transfers.
+            ///
+            /// Burst mode is always enabled for PSRAM transfers.
+            pub external_memory: ExternalBurstConfig,
+
+            /// Enables or disables the burst mode for internal memory transfers.
+            ///
+            /// The burst size is not configurable.
+            pub internal_memory: InternalBurstConfig,
+        }
+
+        impl BurstConfig {
+            /// The default burst mode configuration.
+            pub const DEFAULT: Self = Self {
+                external_memory: ExternalBurstConfig::DEFAULT,
+                internal_memory: InternalBurstConfig::DEFAULT,
+            };
+        }
+
+        impl Default for BurstConfig {
+            fn default() -> Self {
+                Self::DEFAULT
+            }
+        }
+
+        impl From<InternalBurstConfig> for BurstConfig {
+            fn from(internal_memory: InternalBurstConfig) -> Self {
+                Self {
+                    external_memory: ExternalBurstConfig::DEFAULT,
+                    internal_memory,
+                }
+            }
+        }
+
+        impl From<ExternalBurstConfig> for BurstConfig {
+            fn from(external_memory: ExternalBurstConfig) -> Self {
+                Self {
+                    external_memory,
+                    internal_memory: InternalBurstConfig::DEFAULT,
+                }
+            }
+        }
+    } else {
+        /// Burst transfer configuration.
+        #[derive(Clone, Copy, PartialEq, Eq, Debug)]
+        #[cfg_attr(feature = "defmt", derive(defmt::Format))]
+        pub enum BurstConfig {
+            /// Burst mode is disabled.
+            Disabled,
+
+            /// Burst mode is enabled.
+            Enabled,
+        }
+
+        impl BurstConfig {
+            /// The default burst mode configuration.
+            pub const DEFAULT: Self = Self::Disabled;
+        }
+
+        impl Default for BurstConfig {
+            fn default() -> Self {
+                Self::DEFAULT
+            }
+        }
+
+        type InternalBurstConfig = BurstConfig;
+    }
+}
+
+#[cfg(psram_dma)]
+impl ExternalBurstConfig {
+    const fn min_psram_alignment(self, direction: TransferDirection) -> usize {
+        // S2 TRM: Specifically, size and buffer address pointer in receive descriptors
+        // should be 16-byte, 32-byte or 64-byte aligned. For data frame whose
+        // length is not a multiple of 16 bytes, 32 bytes, or 64 bytes, EDMA adds
+        // padding bytes to the end.
+
+        // S3 TRM: Size and Address for IN transfers must be block aligned. For receive
+        // descriptors, if the data length received are not aligned with block size,
+        // GDMA will pad the data received with 0 until they are aligned to
+        // initiate burst transfer. You can read the length field in receive descriptors
+        // to obtain the length of valid data received
+        if matches!(direction, TransferDirection::In) {
+            self as usize
+        } else {
+            // S2 TRM: Size, length and buffer address pointer in transmit descriptors are
+            // not necessarily aligned with block size.
+
+            // S3 TRM: Size, length, and buffer address pointer in transmit descriptors do
+            // not need to be aligned.
+            1
+        }
+    }
+}
+
+impl InternalBurstConfig {
+    pub(super) const fn is_burst_enabled(self) -> bool {
+        !matches!(self, Self::Disabled)
+    }
+
+    // Size and address alignment as those come in pairs on current hardware.
+    const fn min_dram_alignment(self, direction: TransferDirection) -> usize {
+        if matches!(direction, TransferDirection::In) {
+            // NOTE(danielb): commenting this check is incorrect as per TRM, but works.
+            //                we'll need to restore this once peripherals can read a
+            //                different amount of data than what is configured in the
+            //                buffer.
+            // if cfg!(esp32) {
+            //     // NOTE: The size must be word-aligned.
+            //     // NOTE: The buffer address must be word-aligned
+            //     4
+            // }
+            if self.is_burst_enabled() {
+                // As described in "Accessing Internal Memory" paragraphs in the various TRMs.
+                4
+            } else {
+                1
+            }
+        } else {
+            // OUT transfers have no alignment requirements, except for ESP32, which is
+            // described below.
+            if cfg!(esp32) {
+                // SPI DMA: Burst transmission is supported. The data size for
+                // a single transfer must be four bytes aligned.
+                // I2S DMA: Burst transfer is supported. However, unlike the
+                // SPI DMA channels, the data size for a single transfer is
+                // one word, or four bytes.
+                4
+            } else {
+                1
+            }
+        }
+    }
+}
+
+const fn max(a: usize, b: usize) -> usize {
+    if a > b {
+        a
+    } else {
+        b
+    }
 }
 
 impl BurstConfig {
-    pub(super) fn is_burst_enabled(self) -> bool {
-        !matches!(self, Self::Disabled)
+    delegate::delegate! {
+        #[cfg(psram_dma)]
+        to self.internal_memory {
+            pub(super) const fn min_dram_alignment(self, direction: TransferDirection) -> usize;
+            pub(super) fn is_burst_enabled(self) -> bool;
+        }
+    }
+
+    /// Calculates an alignment that is compatible with the current burst
+    /// configuration.
+    ///
+    /// This is an over-estimation so that Descriptors can be safely used with
+    /// any DMA channel in any direction.
+    pub const fn min_compatible_alignment(self) -> usize {
+        let in_alignment = self.min_dram_alignment(TransferDirection::In);
+        let out_alignment = self.min_dram_alignment(TransferDirection::Out);
+        let alignment = max(in_alignment, out_alignment);
+
+        #[cfg(psram_dma)]
+        let alignment = max(alignment, self.external_memory as usize);
+
+        alignment
+    }
+
+    const fn chunk_size_for_alignment(alignment: usize) -> usize {
+        // DMA descriptors have a 12-bit field for the size/length of the buffer they
+        // point at. As there is no such thing as 0-byte alignment, this means the
+        // maximum size is 4095 bytes.
+        4096 - alignment
+    }
+
+    /// Calculates a chunk size that is compatible with the current burst
+    /// configuration's alignment requirements.
+    ///
+    /// This is an over-estimation so that Descriptors can be safely used with
+    /// any DMA channel in any direction.
+    pub const fn max_compatible_chunk_size(self) -> usize {
+        Self::chunk_size_for_alignment(self.min_compatible_alignment())
+    }
+
+    fn min_alignment(self, _buffer: &[u8], direction: TransferDirection) -> usize {
+        let alignment = self.min_dram_alignment(direction);
+
+        cfg_if::cfg_if! {
+            if #[cfg(psram_dma)] {
+                let mut alignment = alignment;
+                if is_valid_psram_address(_buffer.as_ptr() as usize) {
+                    alignment = max(alignment, self.external_memory.min_psram_alignment(direction));
+                }
+            }
+        }
+
+        alignment
+    }
+
+    // Note: this function ignores address alignment as we assume the buffers are
+    // aligned.
+    fn max_chunk_size_for(self, buffer: &[u8], direction: TransferDirection) -> usize {
+        Self::chunk_size_for_alignment(self.min_alignment(buffer, direction))
+    }
+
+    fn ensure_buffer_aligned(
+        self,
+        buffer: &[u8],
+        direction: TransferDirection,
+    ) -> Result<(), DmaAlignmentError> {
+        let alignment = self.min_alignment(buffer, direction);
+        if buffer.as_ptr() as usize % alignment != 0 {
+            return Err(DmaAlignmentError::Address);
+        }
+
+        // NB: the TRMs suggest that buffer length don't need to be aligned, but
+        // for IN transfers, we configure the DMA descriptors' size field, which needs
+        // to be aligned.
+        if direction == TransferDirection::In && buffer.len() % alignment != 0 {
+            return Err(DmaAlignmentError::Size);
+        }
+
+        Ok(())
+    }
+
+    fn ensure_buffer_compatible(
+        self,
+        buffer: &[u8],
+        direction: TransferDirection,
+    ) -> Result<(), DmaBufError> {
+        // buffer can be either DRAM or PSRAM (if supported)
+        let is_in_dram = is_slice_in_dram(buffer);
+        let is_in_psram = cfg!(psram_dma) && is_slice_in_psram(buffer);
+        if !(is_in_dram || is_in_psram) {
+            return Err(DmaBufError::UnsupportedMemoryRegion);
+        }
+
+        self.ensure_buffer_aligned(buffer, direction)?;
+
+        Ok(())
     }
 }
 
@@ -36,6 +352,8 @@ pub enum TransferDirection {
 }
 
 /// Holds all the information needed to configure a DMA channel for a transfer.
+#[derive(PartialEq, Eq, Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct Preparation {
     /// The descriptor the DMA will start from.
     pub start: *mut DmaDescriptor,
@@ -43,29 +361,18 @@ pub struct Preparation {
     /// The direction of the DMA transfer.
     pub direction: TransferDirection,
 
-    /// Block size for PSRAM transfers.
-    ///
-    /// If the buffer is in PSRAM, the implementation must ensure the following:
-    ///
-    /// - The implementation of the buffer must provide a non-`None` block size.
-    /// - For [`TransferDirection::In`] transfers, the implementation of the
-    ///   buffer must invalidate the cache that contains the buffer before the
-    ///   DMA starts.
-    /// - For [`TransferDirection::Out`] transfers, the implementation of the
-    ///   buffer must write back the cache that contains the buffer before the
-    ///   DMA starts.
-    #[cfg(esp32s3)]
-    pub external_memory_block_size: Option<DmaBufBlkSize>,
+    /// Must be `true` if any of the DMA descriptors contain data in PSRAM.
+    #[cfg(psram_dma)]
+    pub accesses_psram: bool,
 
     /// Configures the DMA to transfer data in bursts.
     ///
-    /// The implementation of the buffer must ensure that burst mode is only
-    /// enabled when alignment requirements are met.
+    /// The implementation of the buffer must ensure that buffer size
+    /// and alignment in each descriptor is compatible with the burst
+    /// transfer configuration.
     ///
-    /// There are no additional alignment requirements for
-    /// [`TransferDirection::Out`] burst transfers, but
-    /// [`TransferDirection::In`] transfers require all descriptors to have
-    /// buffer pointers and sizes that are a multiple of 4 (word aligned).
+    /// For details on alignment requirements, refer to your chip's
+    #[doc = crate::trm_markdown_link!()]
     pub burst_transfer: BurstConfig,
 
     /// Configures the "check owner" feature of the DMA channel.
@@ -157,32 +464,6 @@ pub unsafe trait DmaRxBuffer {
 /// descriptors/buffers.
 pub struct BufView<T>(T);
 
-/// Error returned from Dma[Rx|Tx|RxTx]Buf operations.
-#[derive(Debug, PartialEq, Clone, Copy)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum DmaBufError {
-    /// More descriptors are needed for the buffer size
-    InsufficientDescriptors,
-    /// Descriptors or buffers are not located in a supported memory region
-    UnsupportedMemoryRegion,
-    /// Buffer is not aligned to the required size
-    InvalidAlignment,
-    /// Invalid chunk size: must be > 0 and <= 4095
-    InvalidChunkSize,
-}
-
-/// DMA buffer alignments
-#[derive(Debug, Clone, Copy, PartialEq)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum DmaBufBlkSize {
-    /// 16 bytes
-    Size16 = 16,
-    /// 32 bytes
-    Size32 = 32,
-    /// 64 bytes
-    Size64 = 64,
-}
-
 /// DMA transmit buffer
 ///
 /// This is a contiguous buffer linked together by DMA descriptors of length
@@ -193,14 +474,15 @@ pub enum DmaBufBlkSize {
 pub struct DmaTxBuf {
     descriptors: DescriptorSet<'static>,
     buffer: &'static mut [u8],
-    block_size: Option<DmaBufBlkSize>,
+    burst: BurstConfig,
 }
 
 impl DmaTxBuf {
     /// Creates a new [DmaTxBuf] from some descriptors and a buffer.
     ///
     /// There must be enough descriptors for the provided buffer.
-    /// Each descriptor can handle 4092 bytes worth of buffer.
+    /// Depending on alignment requirements, each descriptor can handle at most
+    /// 4095 bytes worth of buffer.
     ///
     /// Both the descriptors and buffer must be in DMA-capable memory.
     /// Only DRAM is supported for descriptors.
@@ -208,79 +490,55 @@ impl DmaTxBuf {
         descriptors: &'static mut [DmaDescriptor],
         buffer: &'static mut [u8],
     ) -> Result<Self, DmaBufError> {
-        Self::new_with_block_size(descriptors, buffer, None)
-    }
-
-    /// Compute max chunk size based on block size
-    pub const fn compute_chunk_size(block_size: Option<DmaBufBlkSize>) -> usize {
-        max_chunk_size(block_size)
-    }
-
-    /// Compute the number of descriptors required for a given block size and
-    /// buffer size
-    pub const fn compute_descriptor_count(
-        buffer_size: usize,
-        block_size: Option<DmaBufBlkSize>,
-    ) -> usize {
-        descriptor_count(buffer_size, Self::compute_chunk_size(block_size), false)
+        Self::new_with_config(descriptors, buffer, BurstConfig::default())
     }
 
     /// Creates a new [DmaTxBuf] from some descriptors and a buffer.
     ///
     /// There must be enough descriptors for the provided buffer.
-    /// Each descriptor can handle at most 4095 bytes worth of buffer.
-    /// Optionally, a block size can be provided for PSRAM & Burst transfers.
+    /// Depending on alignment requirements, each descriptor can handle at most
+    /// 4095 bytes worth of buffer.
     ///
     /// Both the descriptors and buffer must be in DMA-capable memory.
     /// Only DRAM is supported for descriptors.
-    pub fn new_with_block_size(
+    pub fn new_with_config(
         descriptors: &'static mut [DmaDescriptor],
         buffer: &'static mut [u8],
-        block_size: Option<DmaBufBlkSize>,
+        config: impl Into<BurstConfig>,
     ) -> Result<Self, DmaBufError> {
-        cfg_if::cfg_if! {
-            if #[cfg(esp32s3)] {
-                // buffer can be either DRAM or PSRAM (if supported)
-                if !is_slice_in_dram(buffer) && !is_slice_in_psram(buffer) {
-                    return Err(DmaBufError::UnsupportedMemoryRegion);
-                }
-                // if its PSRAM, the block_size/alignment must be specified
-                if is_slice_in_psram(buffer) && block_size.is_none() {
-                    return Err(DmaBufError::InvalidAlignment);
-                }
-            } else {
-                #[cfg(any(esp32,esp32s2))]
-                if buffer.len() % 4 != 0 && buffer.as_ptr() as usize % 4 != 0 {
-                    // ESP32 requires word alignment for DMA buffers.
-                    // ESP32-S2 technically supports byte-aligned DMA buffers, but the
-                    // transfer ends up writing out of bounds if the buffer's length
-                    // is 2 or 3 (mod 4).
-                    return Err(DmaBufError::InvalidAlignment);
-                }
-                // buffer can only be DRAM
-                if !is_slice_in_dram(buffer) {
-                    return Err(DmaBufError::UnsupportedMemoryRegion);
-                }
-            }
-        }
-
-        let block_size = if is_slice_in_dram(buffer) {
-            // no need for block size if the buffer is in DRAM
-            None
-        } else {
-            block_size
-        };
         let mut buf = Self {
             descriptors: DescriptorSet::new(descriptors)?,
             buffer,
-            block_size,
+            burst: BurstConfig::default(),
         };
 
-        buf.descriptors
-            .link_with_buffer(buf.buffer, max_chunk_size(block_size))?;
-        buf.set_length(buf.capacity());
+        let capacity = buf.capacity();
+        buf.configure(config, capacity)?;
 
         Ok(buf)
+    }
+
+    fn configure(
+        &mut self,
+        burst: impl Into<BurstConfig>,
+        length: usize,
+    ) -> Result<(), DmaBufError> {
+        let burst = burst.into();
+        self.set_length_fallible(length, burst)?;
+
+        self.descriptors.link_with_buffer(
+            self.buffer,
+            burst.max_chunk_size_for(self.buffer, TransferDirection::Out),
+        )?;
+
+        self.burst = burst;
+        Ok(())
+    }
+
+    /// Configures the DMA to use burst transfers to access this buffer.
+    pub fn set_burst_config(&mut self, burst: BurstConfig) -> Result<(), DmaBufError> {
+        let len = self.len();
+        self.configure(burst, len)
     }
 
     /// Consume the buf, returning the descriptors and buffer.
@@ -302,17 +560,25 @@ impl DmaTxBuf {
             .sum::<usize>()
     }
 
+    fn set_length_fallible(&mut self, len: usize, burst: BurstConfig) -> Result<(), DmaBufError> {
+        if len > self.capacity() {
+            return Err(DmaBufError::BufferTooSmall);
+        }
+        burst.ensure_buffer_compatible(&self.buffer[..len], TransferDirection::Out)?;
+
+        self.descriptors.set_tx_length(
+            len,
+            burst.max_chunk_size_for(self.buffer, TransferDirection::Out),
+        )
+    }
+
     /// Reset the descriptors to only transmit `len` amount of bytes from this
     /// buf.
     ///
     /// The number of bytes in data must be less than or equal to the buffer
     /// size.
     pub fn set_length(&mut self, len: usize) {
-        assert!(len <= self.buffer.len());
-
-        unwrap!(self
-            .descriptors
-            .set_tx_length(len, max_chunk_size(self.block_size)));
+        unwrap!(self.set_length_fallible(len, self.burst))
     }
 
     /// Fills the TX buffer with the bytes provided in `data` and reset the
@@ -346,23 +612,26 @@ unsafe impl DmaTxBuffer for DmaTxBuf {
             desc.reset_for_tx(desc.next.is_null());
         }
 
-        #[cfg(esp32s3)]
-        if crate::soc::is_valid_psram_address(self.buffer.as_ptr() as usize) {
-            unsafe {
-                crate::soc::cache_writeback_addr(
-                    self.buffer.as_ptr() as u32,
-                    self.buffer.len() as u32,
-                )
-            };
+        cfg_if::cfg_if! {
+            if #[cfg(psram_dma)] {
+                let is_data_in_psram = !is_valid_ram_address(self.buffer.as_ptr() as usize);
+                if is_data_in_psram {
+                    unsafe {
+                        crate::soc::cache_writeback_addr(
+                            self.buffer.as_ptr() as u32,
+                            self.buffer.len() as u32,
+                        )
+                    };
+                }
+            }
         }
 
         Preparation {
             start: self.descriptors.head(),
             direction: TransferDirection::Out,
-            #[cfg(esp32s3)]
-            external_memory_block_size: self.block_size,
-            // TODO: support burst transfers.
-            burst_transfer: BurstConfig::Disabled,
+            #[cfg(psram_dma)]
+            accesses_psram: is_data_in_psram,
+            burst_transfer: self.burst,
             check_owner: None,
         }
     }
@@ -384,6 +653,7 @@ unsafe impl DmaTxBuffer for DmaTxBuf {
 pub struct DmaRxBuf {
     descriptors: DescriptorSet<'static>,
     buffer: &'static mut [u8],
+    burst: BurstConfig,
 }
 
 impl DmaRxBuf {
@@ -398,20 +668,38 @@ impl DmaRxBuf {
         descriptors: &'static mut [DmaDescriptor],
         buffer: &'static mut [u8],
     ) -> Result<Self, DmaBufError> {
-        if !is_slice_in_dram(buffer) {
-            return Err(DmaBufError::UnsupportedMemoryRegion);
-        }
-
         let mut buf = Self {
             descriptors: DescriptorSet::new(descriptors)?,
             buffer,
+            burst: BurstConfig::default(),
         };
 
-        buf.descriptors
-            .link_with_buffer(buf.buffer, max_chunk_size(None))?;
-        buf.set_length(buf.capacity());
+        buf.configure(buf.burst, buf.capacity())?;
 
         Ok(buf)
+    }
+
+    fn configure(
+        &mut self,
+        burst: impl Into<BurstConfig>,
+        length: usize,
+    ) -> Result<(), DmaBufError> {
+        let burst = burst.into();
+        self.set_length_fallible(length, burst)?;
+
+        self.descriptors.link_with_buffer(
+            self.buffer,
+            burst.max_chunk_size_for(self.buffer, TransferDirection::In),
+        )?;
+
+        self.burst = burst;
+        Ok(())
+    }
+
+    /// Configures the DMA to use burst transfers to access this buffer.
+    pub fn set_burst_config(&mut self, burst: BurstConfig) -> Result<(), DmaBufError> {
+        let len = self.len();
+        self.configure(burst, len)
     }
 
     /// Consume the buf, returning the descriptors and buffer.
@@ -434,15 +722,25 @@ impl DmaRxBuf {
             .sum::<usize>()
     }
 
+    fn set_length_fallible(&mut self, len: usize, burst: BurstConfig) -> Result<(), DmaBufError> {
+        if len > self.capacity() {
+            return Err(DmaBufError::BufferTooSmall);
+        }
+        burst.ensure_buffer_compatible(&self.buffer[..len], TransferDirection::In)?;
+
+        self.descriptors.set_rx_length(
+            len,
+            burst.max_chunk_size_for(&self.buffer[..len], TransferDirection::In),
+        )
+    }
+
     /// Reset the descriptors to only receive `len` amount of bytes into this
     /// buf.
     ///
     /// The number of bytes in data must be less than or equal to the buffer
     /// size.
     pub fn set_length(&mut self, len: usize) {
-        assert!(len <= self.buffer.len());
-
-        unwrap!(self.descriptors.set_rx_length(len, max_chunk_size(None)));
+        unwrap!(self.set_length_fallible(len, self.burst));
     }
 
     /// Returns the entire underlying buffer as a slice than can be read.
@@ -503,18 +801,27 @@ unsafe impl DmaRxBuffer for DmaRxBuf {
             desc.reset_for_rx();
         }
 
+        cfg_if::cfg_if! {
+            if #[cfg(psram_dma)] {
+                // Optimization: avoid locking for PSRAM range.
+                let is_data_in_psram = !is_valid_ram_address(self.buffer.as_ptr() as usize);
+                if is_data_in_psram {
+                    unsafe {
+                        crate::soc::cache_invalidate_addr(
+                            self.buffer.as_ptr() as u32,
+                            self.buffer.len() as u32,
+                        )
+                    };
+                }
+            }
+        }
+
         Preparation {
             start: self.descriptors.head(),
             direction: TransferDirection::In,
-
-            // TODO: support external memory access.
-            #[cfg(esp32s3)]
-            external_memory_block_size: None,
-
-            // TODO: DmaRxBuf doesn't currently enforce the alignment requirements required for
-            // bursting. In the future, it could either enforce the alignment or
-            // calculate if the alignment requirements happen to be met.
-            burst_transfer: BurstConfig::Disabled,
+            #[cfg(psram_dma)]
+            accesses_psram: is_data_in_psram,
+            burst_transfer: self.burst,
             check_owner: None,
         }
     }
@@ -538,6 +845,7 @@ pub struct DmaRxTxBuf {
     rx_descriptors: DescriptorSet<'static>,
     tx_descriptors: DescriptorSet<'static>,
     buffer: &'static mut [u8],
+    burst: BurstConfig,
 }
 
 impl DmaRxTxBuf {
@@ -553,22 +861,45 @@ impl DmaRxTxBuf {
         tx_descriptors: &'static mut [DmaDescriptor],
         buffer: &'static mut [u8],
     ) -> Result<Self, DmaBufError> {
-        if !is_slice_in_dram(buffer) {
-            return Err(DmaBufError::UnsupportedMemoryRegion);
-        }
-
         let mut buf = Self {
             rx_descriptors: DescriptorSet::new(rx_descriptors)?,
             tx_descriptors: DescriptorSet::new(tx_descriptors)?,
             buffer,
+            burst: BurstConfig::default(),
         };
-        buf.rx_descriptors
-            .link_with_buffer(buf.buffer, max_chunk_size(None))?;
-        buf.tx_descriptors
-            .link_with_buffer(buf.buffer, max_chunk_size(None))?;
-        buf.set_length(buf.capacity());
+
+        let capacity = buf.capacity();
+        buf.configure(buf.burst, capacity)?;
 
         Ok(buf)
+    }
+
+    fn configure(
+        &mut self,
+        burst: impl Into<BurstConfig>,
+        length: usize,
+    ) -> Result<(), DmaBufError> {
+        let burst = burst.into();
+        self.set_length_fallible(length, burst)?;
+
+        self.rx_descriptors.link_with_buffer(
+            self.buffer,
+            burst.max_chunk_size_for(self.buffer, TransferDirection::In),
+        )?;
+        self.tx_descriptors.link_with_buffer(
+            self.buffer,
+            burst.max_chunk_size_for(self.buffer, TransferDirection::Out),
+        )?;
+
+        self.burst = burst;
+
+        Ok(())
+    }
+
+    /// Configures the DMA to use burst transfers to access this buffer.
+    pub fn set_burst_config(&mut self, burst: BurstConfig) -> Result<(), DmaBufError> {
+        let len = self.len();
+        self.configure(burst, len)
     }
 
     /// Consume the buf, returning the rx descriptors, tx descriptors and
@@ -611,15 +942,31 @@ impl DmaRxTxBuf {
         self.buffer
     }
 
+    fn set_length_fallible(&mut self, len: usize, burst: BurstConfig) -> Result<(), DmaBufError> {
+        if len > self.capacity() {
+            return Err(DmaBufError::BufferTooSmall);
+        }
+        burst.ensure_buffer_compatible(&self.buffer[..len], TransferDirection::In)?;
+        burst.ensure_buffer_compatible(&self.buffer[..len], TransferDirection::Out)?;
+
+        self.rx_descriptors.set_rx_length(
+            len,
+            burst.max_chunk_size_for(self.buffer, TransferDirection::In),
+        )?;
+        self.tx_descriptors.set_tx_length(
+            len,
+            burst.max_chunk_size_for(self.buffer, TransferDirection::Out),
+        )?;
+
+        Ok(())
+    }
+
     /// Reset the descriptors to only transmit/receive `len` amount of bytes
     /// with this buf.
     ///
     /// `len` must be less than or equal to the buffer size.
     pub fn set_length(&mut self, len: usize) {
-        assert!(len <= self.buffer.len());
-
-        unwrap!(self.rx_descriptors.set_rx_length(len, max_chunk_size(None)));
-        unwrap!(self.tx_descriptors.set_tx_length(len, max_chunk_size(None)));
+        unwrap!(self.set_length_fallible(len, self.burst));
     }
 }
 
@@ -633,16 +980,27 @@ unsafe impl DmaTxBuffer for DmaRxTxBuf {
             desc.reset_for_tx(desc.next.is_null());
         }
 
+        cfg_if::cfg_if! {
+            if #[cfg(psram_dma)] {
+                // Optimization: avoid locking for PSRAM range.
+                let is_data_in_psram = !is_valid_ram_address(self.buffer.as_ptr() as usize);
+                if is_data_in_psram {
+                    unsafe {
+                        crate::soc::cache_writeback_addr(
+                            self.buffer.as_ptr() as u32,
+                            self.buffer.len() as u32,
+                        )
+                    };
+                }
+            }
+        }
+
         Preparation {
             start: self.tx_descriptors.head(),
             direction: TransferDirection::Out,
-
-            // TODO: support external memory access.
-            #[cfg(esp32s3)]
-            external_memory_block_size: None,
-
-            // TODO: This is TX, the DMA channel is free to do a burst transfer.
-            burst_transfer: BurstConfig::Disabled,
+            #[cfg(psram_dma)]
+            accesses_psram: is_data_in_psram,
+            burst_transfer: self.burst,
             check_owner: None,
         }
     }
@@ -664,17 +1022,27 @@ unsafe impl DmaRxBuffer for DmaRxTxBuf {
             desc.reset_for_rx();
         }
 
+        cfg_if::cfg_if! {
+            if #[cfg(psram_dma)] {
+                // Optimization: avoid locking for PSRAM range.
+                let is_data_in_psram = !is_valid_ram_address(self.buffer.as_ptr() as usize);
+                if is_data_in_psram {
+                    unsafe {
+                        crate::soc::cache_invalidate_addr(
+                            self.buffer.as_ptr() as u32,
+                            self.buffer.len() as u32,
+                        )
+                    };
+                }
+            }
+        }
+
         Preparation {
             start: self.rx_descriptors.head(),
             direction: TransferDirection::In,
-
-            // TODO: support external memory access.
-            #[cfg(esp32s3)]
-            external_memory_block_size: None,
-
-            // TODO: DmaRxTxBuf doesn't currently enforce the alignment requirements required for
-            // bursting.
-            burst_transfer: BurstConfig::Disabled,
+            #[cfg(psram_dma)]
+            accesses_psram: is_data_in_psram,
+            burst_transfer: self.burst,
             check_owner: None,
         }
     }
@@ -731,6 +1099,7 @@ unsafe impl DmaRxBuffer for DmaRxTxBuf {
 pub struct DmaRxStreamBuf {
     descriptors: &'static mut [DmaDescriptor],
     buffer: &'static mut [u8],
+    burst: BurstConfig,
 }
 
 impl DmaRxStreamBuf {
@@ -789,6 +1158,7 @@ impl DmaRxStreamBuf {
         Ok(Self {
             descriptors,
             buffer,
+            burst: BurstConfig::default(),
         })
     }
 
@@ -808,14 +1178,9 @@ unsafe impl DmaRxBuffer for DmaRxStreamBuf {
         Preparation {
             start: self.descriptors.as_mut_ptr(),
             direction: TransferDirection::In,
-
-            // TODO: support external memory access.
-            #[cfg(esp32s3)]
-            external_memory_block_size: None,
-
-            // TODO: DmaRxStreamBuf doesn't currently enforce the alignment requirements required
-            // for bursting.
-            burst_transfer: BurstConfig::Disabled,
+            #[cfg(psram_dma)]
+            accesses_psram: false,
+            burst_transfer: self.burst,
 
             // Whilst we give ownership of the descriptors the DMA, the correctness of this buffer
             // implementation doesn't rely on the DMA checking for descriptor ownership.
@@ -1022,9 +1387,9 @@ unsafe impl DmaTxBuffer for EmptyBuf {
         Preparation {
             start: unsafe { core::ptr::addr_of_mut!(EMPTY).cast() },
             direction: TransferDirection::Out,
-            #[cfg(esp32s3)]
-            external_memory_block_size: None,
-            burst_transfer: BurstConfig::Disabled,
+            #[cfg(psram_dma)]
+            accesses_psram: false,
+            burst_transfer: BurstConfig::default(),
 
             // As we don't give ownership of the descriptor to the DMA, it's important that the DMA
             // channel does *NOT* check for ownership, otherwise the channel will return an error.
@@ -1049,9 +1414,9 @@ unsafe impl DmaRxBuffer for EmptyBuf {
         Preparation {
             start: unsafe { core::ptr::addr_of_mut!(EMPTY).cast() },
             direction: TransferDirection::In,
-            #[cfg(esp32s3)]
-            external_memory_block_size: None,
-            burst_transfer: BurstConfig::Disabled,
+            #[cfg(psram_dma)]
+            accesses_psram: false,
+            burst_transfer: BurstConfig::default(),
 
             // As we don't give ownership of the descriptor to the DMA, it's important that the DMA
             // channel does *NOT* check for ownership, otherwise the channel will return an error.
@@ -1096,7 +1461,8 @@ impl DmaLoopBuf {
             return Err(DmaBufError::UnsupportedMemoryRegion);
         }
 
-        if buffer.len() > max_chunk_size(None) {
+        if buffer.len() > BurstConfig::default().max_chunk_size_for(buffer, TransferDirection::Out)
+        {
             return Err(DmaBufError::InsufficientDescriptors);
         }
 
@@ -1122,11 +1488,10 @@ unsafe impl DmaTxBuffer for DmaLoopBuf {
     fn prepare(&mut self) -> Preparation {
         Preparation {
             start: self.descriptor,
+            #[cfg(psram_dma)]
+            accesses_psram: false,
             direction: TransferDirection::Out,
-            // TODO: support external memory access.
-            #[cfg(esp32s3)]
-            external_memory_block_size: None,
-            burst_transfer: BurstConfig::Disabled,
+            burst_transfer: BurstConfig::default(),
             // The DMA must not check the owner bit, as it is never set.
             check_owner: Some(false),
         }