+++ esp-hal/src/dma/mod.rs
@@ -64,7 +64,7 @@ use crate::{
     interrupt::InterruptHandler,
     peripheral::{Peripheral, PeripheralRef},
     peripherals::Interrupt,
-    soc::is_slice_in_dram,
+    soc::{is_slice_in_dram, is_valid_memory_address, is_valid_ram_address},
     system,
     Async,
     Blocking,
@@ -675,6 +675,14 @@ macro_rules! dma_buffers_impl {
             )
         }
     }};
+
+    ($size:expr, is_circular = $circular:tt) => {
+        $crate::dma_buffers_impl!(
+            $size,
+            $crate::dma::BurstConfig::DEFAULT.max_compatible_chunk_size(),
+            is_circular = $circular
+        );
+    };
 }
 
 #[doc(hidden)]
@@ -722,7 +730,6 @@ macro_rules! dma_descriptor_count {
 /// ```rust,no_run
 #[doc = crate::before_snippet!()]
 /// use esp_hal::dma_tx_buffer;
-/// use esp_hal::dma::DmaBufBlkSize;
 ///
 /// let tx_buf = dma_tx_buffer!(32000);
 /// # }
@@ -730,11 +737,7 @@ macro_rules! dma_descriptor_count {
 #[macro_export]
 macro_rules! dma_tx_buffer {
     ($tx_size:expr) => {{
-        let (tx_buffer, tx_descriptors) = $crate::dma_buffers_impl!(
-            $tx_size,
-            $crate::dma::DmaTxBuf::compute_chunk_size(None),
-            is_circular = false
-        );
+        let (tx_buffer, tx_descriptors) = $crate::dma_buffers_impl!($tx_size, is_circular = false);
 
         $crate::dma::DmaTxBuf::new(tx_descriptors, tx_buffer)
     }};
@@ -794,11 +797,11 @@ macro_rules! dma_loop_buffer {
 }
 
 /// DMA Errors
-#[derive(Debug, Clone, Copy, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum DmaError {
     /// The alignment of data is invalid
-    InvalidAlignment,
+    InvalidAlignment(DmaAlignmentError),
     /// More descriptors are needed for the buffer size
     OutOfDescriptors,
     /// DescriptorError the DMA rejected the descriptor configuration. This
@@ -824,8 +827,9 @@ impl From<DmaBufError> for DmaError {
         match error {
             DmaBufError::InsufficientDescriptors => DmaError::OutOfDescriptors,
             DmaBufError::UnsupportedMemoryRegion => DmaError::UnsupportedMemoryRegion,
-            DmaBufError::InvalidAlignment => DmaError::InvalidAlignment,
+            DmaBufError::InvalidAlignment(err) => DmaError::InvalidAlignment(err),
             DmaBufError::InvalidChunkSize => DmaError::InvalidChunkSize,
+            DmaBufError::BufferTooSmall => DmaError::BufferTooSmall,
         }
     }
 }
@@ -1016,10 +1020,10 @@ impl DescriptorChain {
         len: usize,
         prepare_descriptor: impl Fn(&mut DmaDescriptor, usize),
     ) -> Result<(), DmaError> {
-        if !crate::soc::is_valid_ram_address(self.first() as usize)
-            || !crate::soc::is_valid_ram_address(self.last() as usize)
-            || !crate::soc::is_valid_memory_address(data as usize)
-            || !crate::soc::is_valid_memory_address(unsafe { data.add(len) } as usize)
+        if !is_valid_ram_address(self.first() as usize)
+            || !is_valid_ram_address(self.last() as usize)
+            || !is_valid_memory_address(data as usize)
+            || !is_valid_memory_address(unsafe { data.add(len) } as usize)
         {
             return Err(DmaError::UnsupportedMemoryRegion);
         }
@@ -1064,17 +1068,6 @@ pub const fn descriptor_count(buffer_size: usize, chunk_size: usize, is_circular
     }
 
     buffer_size.div_ceil(chunk_size)
-}
-
-/// Compute max chunk size based on block size.
-const fn max_chunk_size(block_size: Option<DmaBufBlkSize>) -> usize {
-    match block_size {
-        Some(size) => 4096 - size as usize,
-        #[cfg(esp32)]
-        None => 4092, // esp32 requires 4 byte alignment
-        #[cfg(not(esp32))]
-        None => 4095,
-    }
 }
 
 #[derive(Debug)]
@@ -1272,6 +1265,7 @@ impl<'a> DescriptorSet<'a> {
 }
 
 /// Block size for transfers to/from PSRAM
+#[cfg(psram_dma)]
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub enum DmaExtMemBKSize {
     /// External memory block size of 16 bytes.
@@ -1282,12 +1276,13 @@ pub enum DmaExtMemBKSize {
     Size64 = 2,
 }
 
-impl From<DmaBufBlkSize> for DmaExtMemBKSize {
-    fn from(size: DmaBufBlkSize) -> Self {
+#[cfg(psram_dma)]
+impl From<ExternalBurstConfig> for DmaExtMemBKSize {
+    fn from(size: ExternalBurstConfig) -> Self {
         match size {
-            DmaBufBlkSize::Size16 => DmaExtMemBKSize::Size16,
-            DmaBufBlkSize::Size32 => DmaExtMemBKSize::Size32,
-            DmaBufBlkSize::Size64 => DmaExtMemBKSize::Size64,
+            ExternalBurstConfig::Size16 => DmaExtMemBKSize::Size16,
+            ExternalBurstConfig::Size32 => DmaExtMemBKSize::Size32,
+            ExternalBurstConfig::Size64 => DmaExtMemBKSize::Size64,
         }
     }
 }
@@ -1758,9 +1753,6 @@ pub trait Rx: crate::private::Sealed {
 
     fn stop_transfer(&mut self);
 
-    #[cfg(esp32s3)]
-    fn set_ext_mem_block_size(&self, size: DmaExtMemBKSize);
-
     #[cfg(gdma)]
     fn set_mem2mem_mode(&mut self, value: bool);
 
@@ -1916,6 +1908,17 @@ where
     ) -> Result<(), DmaError> {
         debug_assert_eq!(preparation.direction, TransferDirection::In);
 
+        debug!("Preparing RX transfer {:?}", preparation);
+        trace!("First descriptor {:?}", unsafe { &*preparation.start });
+
+        #[cfg(psram_dma)]
+        if preparation.accesses_psram && !self.rx_impl.can_access_psram() {
+            return Err(DmaError::UnsupportedMemoryRegion);
+        }
+
+        #[cfg(psram_dma)]
+        self.rx_impl
+            .set_ext_mem_block_size(preparation.burst_transfer.external_memory.into());
         self.rx_impl.set_burst_mode(preparation.burst_transfer);
         self.rx_impl.set_descr_burst_mode(true);
         self.rx_impl.set_check_owner(preparation.check_owner);
@@ -1950,36 +1953,43 @@ where
         peri: DmaPeripheral,
         chain: &DescriptorChain,
     ) -> Result<(), DmaError> {
-        // For ESP32-S3 we check each descriptor buffer that points to PSRAM for
+        // We check each descriptor buffer that points to PSRAM for
         // alignment and invalidate the cache for that buffer.
         // NOTE: for RX the `buffer` and `size` need to be aligned but the `len` does
         // not. TRM section 3.4.9
         // Note that DmaBuffer implementations are required to do this for us.
-        #[cfg(esp32s3)]
-        for des in chain.descriptors.iter() {
-            // we are forcing the DMA alignment to the cache line size
-            // required when we are using dcache
-            let alignment = crate::soc::cache_get_dcache_line_size() as usize;
-            if crate::soc::is_valid_psram_address(des.buffer as usize) {
-                // both the size and address of the buffer must be aligned
-                if des.buffer as usize % alignment != 0 && des.size() % alignment != 0 {
-                    return Err(DmaError::InvalidAlignment);
+        cfg_if::cfg_if! {
+            if #[cfg(psram_dma)] {
+                let mut uses_psram = false;
+                let psram_range = crate::soc::psram_range();
+                for des in chain.descriptors.iter() {
+                    // we are forcing the DMA alignment to the cache line size
+                    // required when we are using dcache
+                    let alignment = crate::soc::cache_get_dcache_line_size() as usize;
+                    if crate::soc::addr_in_range(des.buffer as usize, psram_range.clone()) {
+                        uses_psram = true;
+                        // both the size and address of the buffer must be aligned
+                        if des.buffer as usize % alignment != 0 {
+                            return Err(DmaError::InvalidAlignment(DmaAlignmentError::Address));
+                        }
+                        if des.size() % alignment != 0 {
+                            return Err(DmaError::InvalidAlignment(DmaAlignmentError::Size));
+                        }
+                        crate::soc::cache_invalidate_addr(des.buffer as u32, des.size() as u32);
+                    }
                 }
-                crate::soc::cache_invalidate_addr(des.buffer as u32, des.size() as u32);
             }
         }
 
-        self.do_prepare(
-            Preparation {
-                start: chain.first().cast_mut(),
-                #[cfg(esp32s3)]
-                external_memory_block_size: None,
-                direction: TransferDirection::In,
-                burst_transfer: BurstConfig::Disabled,
-                check_owner: Some(false),
-            },
-            peri,
-        )
+        let preparation = Preparation {
+            start: chain.first().cast_mut(),
+            direction: TransferDirection::In,
+            #[cfg(psram_dma)]
+            accesses_psram: uses_psram,
+            burst_transfer: BurstConfig::default(),
+            check_owner: Some(false),
+        };
+        self.do_prepare(preparation, peri)
     }
 
     unsafe fn prepare_transfer<BUF: DmaRxBuffer>(
@@ -2007,11 +2017,6 @@ where
 
     fn stop_transfer(&mut self) {
         self.rx_impl.stop()
-    }
-
-    #[cfg(esp32s3)]
-    fn set_ext_mem_block_size(&self, size: DmaExtMemBKSize) {
-        self.rx_impl.set_ext_mem_block_size(size);
     }
 
     #[cfg(gdma)]
@@ -2081,9 +2086,6 @@ pub trait Tx: crate::private::Sealed {
     fn start_transfer(&mut self) -> Result<(), DmaError>;
 
     fn stop_transfer(&mut self);
-
-    #[cfg(esp32s3)]
-    fn set_ext_mem_block_size(&self, size: DmaExtMemBKSize);
 
     fn is_done(&self) -> bool {
         self.pending_out_interrupts()
@@ -2200,11 +2202,17 @@ where
     ) -> Result<(), DmaError> {
         debug_assert_eq!(preparation.direction, TransferDirection::Out);
 
-        #[cfg(esp32s3)]
-        if let Some(block_size) = preparation.external_memory_block_size {
-            self.set_ext_mem_block_size(block_size.into());
+        debug!("Preparing TX transfer {:?}", preparation);
+        trace!("First descriptor {:?}", unsafe { &*preparation.start });
+
+        #[cfg(psram_dma)]
+        if preparation.accesses_psram && !self.tx_impl.can_access_psram() {
+            return Err(DmaError::UnsupportedMemoryRegion);
         }
 
+        #[cfg(psram_dma)]
+        self.tx_impl
+            .set_ext_mem_block_size(preparation.burst_transfer.external_memory.into());
         self.tx_impl.set_burst_mode(preparation.burst_transfer);
         self.tx_impl.set_descr_burst_mode(true);
         self.tx_impl.set_check_owner(preparation.check_owner);
@@ -2241,34 +2249,42 @@ where
     ) -> Result<(), DmaError> {
         // Based on the ESP32-S3 TRM the alignment check is not needed for TX
 
-        // For esp32s3 we check each descriptor buffer that points to PSRAM for
+        // We check each descriptor buffer that points to PSRAM for
         // alignment and writeback the cache for that buffer.
         // Note that DmaBuffer implementations are required to do this for us.
-        #[cfg(esp32s3)]
-        for des in chain.descriptors.iter() {
-            // we are forcing the DMA alignment to the cache line size
-            // required when we are using dcache
-            let alignment = crate::soc::cache_get_dcache_line_size() as usize;
-            if crate::soc::is_valid_psram_address(des.buffer as usize) {
-                // both the size and address of the buffer must be aligned
-                if des.buffer as usize % alignment != 0 && des.size() % alignment != 0 {
-                    return Err(DmaError::InvalidAlignment);
+        #[cfg(psram_dma)]
+        cfg_if::cfg_if! {
+            if #[cfg(psram_dma)] {
+                let mut uses_psram = false;
+                let psram_range = crate::soc::psram_range();
+                for des in chain.descriptors.iter() {
+                    // we are forcing the DMA alignment to the cache line size
+                    // required when we are using dcache
+                    let alignment = crate::soc::cache_get_dcache_line_size() as usize;
+                    if crate::soc::addr_in_range(des.buffer as usize, psram_range.clone()) {
+                        uses_psram = true;
+                        // both the size and address of the buffer must be aligned
+                        if des.buffer as usize % alignment != 0 {
+                            return Err(DmaError::InvalidAlignment(DmaAlignmentError::Address));
+                        }
+                        if des.size() % alignment != 0 {
+                            return Err(DmaError::InvalidAlignment(DmaAlignmentError::Size));
+                        }
+                        crate::soc::cache_writeback_addr(des.buffer as u32, des.size() as u32);
+                    }
                 }
-                crate::soc::cache_writeback_addr(des.buffer as u32, des.size() as u32);
             }
         }
 
-        self.do_prepare(
-            Preparation {
-                start: chain.first().cast_mut(),
-                #[cfg(esp32s3)]
-                external_memory_block_size: None,
-                direction: TransferDirection::Out,
-                burst_transfer: BurstConfig::Disabled,
-                check_owner: Some(false),
-            },
-            peri,
-        )?;
+        let preparation = Preparation {
+            start: chain.first().cast_mut(),
+            direction: TransferDirection::Out,
+            #[cfg(psram_dma)]
+            accesses_psram: uses_psram,
+            burst_transfer: BurstConfig::default(),
+            check_owner: Some(false),
+        };
+        self.do_prepare(preparation, peri)?;
 
         // enable descriptor write back in circular mode
         self.tx_impl
@@ -2302,11 +2318,6 @@ where
 
     fn stop_transfer(&mut self) {
         self.tx_impl.stop()
-    }
-
-    #[cfg(esp32s3)]
-    fn set_ext_mem_block_size(&self, size: DmaExtMemBKSize) {
-        self.tx_impl.set_ext_mem_block_size(size);
     }
 
     fn listen_out(&self, interrupts: impl Into<EnumSet<DmaTxInterrupt>>) {
@@ -2379,11 +2390,14 @@ pub trait RegisterAccess: crate::private::Sealed {
     /// descriptor.
     fn set_check_owner(&self, check_owner: Option<bool>);
 
-    #[cfg(esp32s3)]
+    #[cfg(psram_dma)]
     fn set_ext_mem_block_size(&self, size: DmaExtMemBKSize);
 
     #[cfg(pdma)]
     fn is_compatible_with(&self, peripheral: DmaPeripheral) -> bool;
+
+    #[cfg(psram_dma)]
+    fn can_access_psram(&self) -> bool;
 }
 
 #[doc(hidden)]